Implementing item rating

+++ backend/src/controllers/ordersController.js
@@ -5,16 +5,26 @@ module.exports = {
     create: async (req, res, next) => {
         const {item_id, customer_id, quantity} = req.body;
 
-        const itemSql = 'SELECT `price` FROM `items` WHERE `id` = ?';
-        const [itemResponseObject, sqlError] = await makeSqlQuery(itemSql, [item_id]);
+        let sql = 'SELECT `stock`, `price` FROM `items` WHERE `id` = ?';
+        const [itemResponseObject, sqlError] = await makeSqlQuery(sql, [item_id]);
+
+        if (sqlError) {
+            return next(sqlError);
+        }
+
         if (itemResponseObject.length === 0) {
             return next(new APIError('Item not found', 404));
+        }
+
+        const itemStock = itemResponseObject[0].stock;
+        if (quantity > itemStock) {
+            return next(new APIError('Item stock is too low', 400));
         }
 
         const itemPrice = itemResponseObject[0].price;
         const orderTotal = itemPrice * quantity;
 
-        const sql = 'INSERT INTO `orders` (`item_id`, `customer_id`, `qty`, `price`, `total`) VALUES (?, ?, ?, ?, ?)';
+        sql = 'INSERT INTO `orders` (`item_id`, `customer_id`, `qty`, `price`, `total`) VALUES (?, ?, ?, ?, ?)';
 
         const [responseObject, error] = await makeSqlQuery(sql, [
             item_id,
@@ -29,6 +39,22 @@ module.exports = {
         }
 
         if (responseObject.affectedRows !== 1) {
+            return next(new APIError('something went wrong', 400));
+        }
+
+        sql = 'UPDATE `items` SET `stock` = `stock` - ? WHERE `id` = ?' ;
+        const [responseObjectItem, errorItem] = await makeSqlQuery(sql, [quantity, item_id]);
+        if (errorItem) {
+            return next(errorItem);
+        }
+
+        if (responseObjectItem.affectedRows !== 1) {
+            sql = 'DELETE FROM `orders` WHERE `id` = ?';
+            const [responseObjectOrderDelete, errorOrderDelete] = await makeSqlQuery(sql, [responseObject.insertId]);
+            if (errorOrderDelete) {
+                return next(errorOrderDelete);
+            }
+
             return next(new APIError('something went wrong', 400));
         }
 
@@ -102,7 +128,29 @@ module.exports = {
             return next(new APIError('Access denied!', 401));
         }
 
-        const sql = 'DELETE FROM `orders` WHERE id = ?';
+        let sql = 'SELECT `item_id`, `qty` FROM `orders` WHERE `id` = ?';
+
+        const [orderResponseObject, sqlError] = await makeSqlQuery(sql, [id]);
+
+        if (sqlError) {
+            return next(sqlError);
+        }
+
+        if (orderResponseObject.length === 0) {
+            return next(new APIError('Order not found', 404));
+        }
+
+        sql = 'UPDATE `items` SET `stock` = `stock` + ? WHERE `id` = ?';
+        const [itemResponseObject, itemError] = await makeSqlQuery(sql, [
+            orderResponseObject[0].qty,
+            orderResponseObject[0].item_id,
+        ]);
+
+        if (itemError) {
+            return next(itemError);
+        }
+
+        sql = 'DELETE FROM `orders` WHERE id = ?';
 
         const [responseObject, error] = await makeSqlQuery(sql, [id]);
         if (error) {
